refactor: Create RacingController
- 레이싱 컨트롤러 생성
- 레이싱 머신 역할 분할
