+++ src/main/java/racingcar/RacingController.java
@@ -1,0 +1,13 @@
+package racingcar;
+
+import racingcar.domain.RacingMachine;
+import racingcar.util.InputView;
+
+public class RacingController {
+
+    public static void main(String[] args) {
+        int numberOfCars = InputView.getNumberOfCar();
+        int numberOfRound = InputView.getNumberOfRound();
+        RacingMachine.runMachine(numberOfCars, numberOfRound);
+    }
+}
