+++ src/main/java/racingcar/domain/RacingMachine.java
@@ -1,23 +1,16 @@
-package racingcar;
+package racingcar.domain;
 
-import racingcar.domain.Cars;
-import racingcar.util.InputView;
 import racingcar.util.ResultView;
 
 public class RacingMachine {
     private static final int ROUND_END = 0;
 
-    private static int numberOfCars;
-    private static int numberOfRounds;
     private static Cars cars;
 
-    public static void main(String[] args) {
-        runMachine();
+    private RacingMachine() {
     }
 
-    private static void runMachine() {
-        numberOfCars = InputView.getNumberOfCar();
-        numberOfRounds = InputView.getNumberOfRound();
+    public static void runMachine(int numberOfCars, int numberOfRounds) {
         cars = new Cars(numberOfCars);
         ResultView.startGame();
         while (numberOfRounds-- > ROUND_END) {
